refactor: DynamicTest 내에 assert문을 선언하는 것으로 변경
